+++ src/sqlfmt/rules/common.py
@@ -37,3 +37,18 @@ ALTER_DROP_FUNCTION = r"(alter|drop)\s+function(\s+if\s+exists)?"
 
 CREATE_WAREHOUSE = r"create(\s+or\s+replace)?\s+warehouse(\s+if\s+not\s+exists)?"
 ALTER_WAREHOUSE = r"alter\s+warehouse(\s+if\s+exists)?"
+
+CREATE_CLONABLE = (
+    r"create(\s+or\s+replace)?\s+"
+    + group(
+        r"database",
+        r"schema",
+        r"table",
+        r"stage",
+        r"file\s+format",
+        r"sequence",
+        r"stream",
+        r"task",
+    )
+    + r"(\s+if\s+not\s+exists)?"
+)